requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.19
django-embed-video==1.4.8
djangorestframework==3.14.0
flake8==6.0.0
idna==3.4
mccabe==0.7.0
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
urllib3==1.26.15
